+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Bug Fixes
 
 -   When running multiple queries, Harlequin now activates the results tab for the last query, instead of the first one.
+-   Queries that return duplicate column names are now displayed correctly in the Results Viewer ([textual-fastdatatable/#26](https://github.com/tconbeer/textual-fastdatatable/issues/26)).
 
 ## [1.3.0] - 2023-11-06
 
@@ -16,7 +17,7 @@ All notable changes to this project will be documented in this file.
 
 ### Bug Fixes
 
--   Fixes a crash that could happen when a query returned no records ([#300](https://github.com/tconbeer/textual-fastdatatable/issues/19)).
+-   Fixes a crash that could happen when a query returned no records ([textual-fastdatatable/#19](https://github.com/tconbeer/textual-fastdatatable/issues/19)).
 
 ### Adapter API Changes